PathwaySpace: Spatial Projection of Network Signals along Geodesic Paths

For a given graph containing vertices, edges, and a signal associated with the vertices, the 'PathwaySpace' package performs a convolution operation, which involves a weighted combination of neighboring vertices and their associated signals. The package uses a decay function to project these signals, creating geodesic paths on a 2D-image space. 'PathwaySpace' has various applications, such as visualizing network data in a graphical format that highlights the relationships and signal strengths between vertices. By combining graph theory, signal processing, and visualization, 'PathwaySpace' provides a way of representing graph data on a continuous projection space. Based on methods introduced in Tercan et al. (2025) doi:10.1016/j.xpro.2025.103681 and Ellrott et al. (2025) doi:10.1016/j.ccell.2024.12.002 .

Index

PathwaySpace-class:	An S4 class for signal propagation on pathway spaces.

buildPathwaySpace:	Constructor of PathwaySpace-class objects.

circularProjection:	Creating 2D-landscape images from graph objects.

polarProjection:	Creating 2D-landscape images from graph objects.

silhouetteMapping:	Mapping graph silhouettes on PathwaySpace images.

summitMapping:	Mapping summits on a 2D-landscape image.

getPathwaySpace:	Accessory method for fetching slots from a PathwaySpace object.

plotPathwaySpace:	Plotting 2D-landscape images for the PathwaySpace package.

Further information is available in the vignettes by typing vignette('PathwaySpace'). Documented topics are also available in HTML by typing help.start() and selecting the PathwaySpace package from the menu.

References

Sysbiolab Team (2026). RGraphSpace: A lightweight interface between 'igraph' and 'ggplot2' graphics. R package version 1.4.1 (Doi: 10.32614/CRAN.package.RGraphSpace), https://CRAN.R-project.org/package=RGraphSpace.

See also

Author

Maintainer: Mauro Castro mauro.a.castro@gmail.com (ORCID)

Authors:

Sysbiolab Team

Other contributors:

Victor Apolonio [contributor]

Jonathan Back [contributor]

Lana Querne [contributor]

Vinicius Chagas [contributor]

Bahar Tercan [contributor]