getNearestNode

Retrieves the nearest neighbor for each node from a PathwaySpace object using Euclidean distance.

Usage

getNearestNode(ps)

Arguments

ps

Either a PathwaySpace or GraphSpace object.

Value

A data.frame with columns from, to, and dist, listing each node's nearest neighbor and the Euclidean distance between them.

See also

Examples

# See examples in the PathwaySpace's tutorials:
# https://sysbiolab.github.io/PathwaySpace/